+++ examples/k8s_cluster.py
@@ -69,5 +69,6 @@ if __name__ == "__main__":
     # process knows the address of all other processes in the cluster
     # and their unique process ID. You can address that easily by deploying your
     # dataflow program using Waxctl or installing the Bytewax Helm Chart
-    input_config = ManualInputConfig(input_builder)
-    cluster_main(flow, input_config, output_builder, **parse.proc_env())
+    cluster_main(
+        flow, ManualInputConfig(input_builder), output_builder, **parse.proc_env()
+    )
