+++ examples/recovery_wordcount.py
@@ -56,13 +56,12 @@ flow.capture()
 
 
 if __name__ == "__main__":
-    input_config = ManualInputConfig(input_builder)
     recovery_config = KafkaRecoveryConfig(
         ["localhost:9092"], "bytewax-state", create=True
     )
     spawn_cluster(
         flow,
-        input_config,
+        ManualInputConfig(input_builder),
         output_builder,
         recovery_config=recovery_config,
         **parse.cluster_args()
