+++ pysrc/bytewax/execution.py
@@ -8,7 +8,14 @@ from multiprocess import get_context
 from bytewax.recovery import RecoveryConfig
 from bytewax.inputs import InputConfig
 
-from .bytewax import AdvanceTo, cluster_main, Dataflow, Emit, ManualInputConfig, run_main
+from .bytewax import (
+    AdvanceTo,
+    cluster_main,
+    Dataflow,
+    Emit,
+    ManualInputConfig,
+    run_main,
+)
 
 
 def run(
@@ -64,10 +71,9 @@ def run(
         assert worker_index == 0
         return out.append
 
-    input_config = ManualInputConfig(input_builder)
     run_main(
         flow,
-        input_config,
+        ManualInputConfig(input_builder),
         output_builder,
         recovery_config=recovery_config,
     )
@@ -258,13 +264,12 @@ def run_cluster(
 
         out = man.list()
 
-        input_config = ManualInputConfig(input_builder)
         def output_builder(worker_index, worker_count):
             return out.append
 
         spawn_cluster(
             flow,
-            input_config,
+            ManualInputConfig(input_builder),
             output_builder,
             recovery_config=recovery_config,
             proc_count=proc_count,
